+++ src/cmd/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod anchors;
 pub mod covered;
+pub mod dazzname;
 pub mod dep;
 pub mod ena;
 pub mod merge;